Add customer address form, update mutation queries
