--- app/graphql/customer-account/customer-address-mutations.ts
@@ -1,58 +1,0 @@
-// NOTE: https://shopify.dev/docs/api/customer/latest/mutations/customerAddressUpdate
-export const UPDATE_ADDRESS_MUTATION = `#graphql
-  mutation customerAddressUpdate(
-    $address: CustomerAddressInput!
-    $addressId: ID!
-    $defaultAddress: Boolean
- ) {
-    customerAddressUpdate(
-      address: $address
-      addressId: $addressId
-      defaultAddress: $defaultAddress
-    ) {
-      userErrors {
-        code
-        field
-        message
-      }
-    }
-  }
-` as const;
-
-// NOTE: https://shopify.dev/docs/api/customer/latest/mutations/customerAddressDelete
-export const DELETE_ADDRESS_MUTATION = `#graphql
-  mutation customerAddressDelete(
-    $addressId: ID!,
-  ) {
-    customerAddressDelete(addressId: $addressId) {
-      deletedAddressId
-      userErrors {
-        code
-        field
-        message
-      }
-    }
-  }
-` as const;
-
-// NOTE: https://shopify.dev/docs/api/customer/latest/mutations/customerAddressCreate
-export const CREATE_ADDRESS_MUTATION = `#graphql
-  mutation customerAddressCreate(
-    $address: CustomerAddressInput!
-    $defaultAddress: Boolean
-  ) {
-    customerAddressCreate(
-      address: $address
-      defaultAddress: $defaultAddress
-    ) {
-      customerAddress {
-        id
-      }
-      userErrors {
-        code
-        field
-        message
-      }
-    }
-  }
-` as const;
